backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
python-dotenv
bcrypt